WR#341276: Update for allowing the first branched reponse to be required.
This does not set the required field in {user_info_field} as at any point, if this was not filled out then the users profile would never pass the user_not_fully_setup check.

This hooks into 'after_require_login' to determine if the new user profile fields are required or not.